Version_2/pyproject.toml
[project]
name = "version-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.6.0",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
]
